Fix: media player not working after device rename (#51)
This PR fixes issue where renaming device would cause the media player entity not to be updated in Home Assistant.
Combined with additional change in 2.2.0 version of HASS.Agent should ensure the media player works properly.

Potentially related issues (to be verified after this PR is published in one of the beta versions of the integration):

Bug: Can't get notifications or mediaplayer to work. Something wrong after reinstall with device id logic #34

    Bug: Media Playera is Unavailable after device rename #27

Note: the media player entity ID in Home Assistant will remain the same, this means that history and configured card/dashboards will not require modification. The notify service calls WILL need to be updated though.